Refactor material buffers to prep for realistic materials

+++ src/main/java/grondag/hard_science/library/varia/Useful.java
@@ -487,7 +487,7 @@ public class Useful
         return (intersectionCount & 0x1) == 1;
     }
 
-    public static double clamp(double min, double max, double val)
+    public static long clamp(long min, long max, long val)
     {
         return val < min ? min : val > max ? max : val;
     }
